RetroPlayer: Rename windows/ folder to guiwindows/
Original idea credit vel0city.
