+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreen.cpp
