+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreen.h
