+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreenText.cpp
