+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreenText.h
